backend/src/question/requirements.txt
fastapi
uvicorn
supabase
pydantic
redis
pika
python-consul
prometheus-fastapi-instrumentator
python-dotenv